+++ error.c
@@ -245,7 +245,7 @@ noreturn static pic_value
 pic_error_error(pic_state *pic)
 {
   const char *str;
-  int argc;
+  size_t argc;
   pic_value *argv;
 
   pic_get_args(pic, "z*", &str, &argc, &argv);
@@ -259,7 +259,7 @@ pic_error_make_error_object(pic_state *pic)
   struct pic_error *e;
   pic_sym type;
   pic_str *msg;
-  int argc;
+  size_t argc;
   pic_value *argv;
 
   pic_get_args(pic, "ms*", &type, &msg, &argc, &argv);
