+++ symbol.c
@@ -108,7 +108,7 @@ pic_symbol_symbol_p(pic_state *pic)
 static pic_value
 pic_symbol_symbol_eq_p(pic_state *pic)
 {
-  int argc, i;
+  size_t argc, i;
   pic_value *argv;
 
   pic_get_args(pic, "*", &argc, &argv);
